Fixes missing bandcamp location.
This closes #6280.
This closes #6281.
This closes #6282.
This closes #6283.
This closes #6284.
This closes #6285.
This closes #6286.
This closes #6287.
This closes #6288.
This closes #6289.
This closes #6290.
This closes #6291.
This closes #6292.
This closes #6293.
This closes #6294.
This closes #6295.
This closes #6296.
This closes #6297.
This closes #6298.
This closes #6299.
This closes #6300.
This closes #6301.
This closes #6302.
This closes #6303.
This closes #6304.
This closes #6305.
This closes #6306.
This closes #6307.
This closes #6308.
This closes #6309.
This closes #6310.
This closes #6311.
This closes #6312.
This closes #6313.
This closes #6314.
This closes #6315.
This closes #6316.
This closes #6317.
This closes #6318.
This closes #6319.
This closes #6320.